- I noticed was that the fix triangles weren't centered properly on the fix position.  (I "fixed" that problem) :-)
- I fixed a problem in how atlas figured out magnetic offset of the
vor's (we had lat, and lon switched ... probably my fault.)

- As I mentioned before, I created a "proprietary" sentence to
communicate the nav radio settings to atlas.  I added a --atlas output
format to flightgear and added the corresponding code to the Atlas
program to parse the new sentence.

Issue #1: how best to communicate the radio settings inside Atlas from
the parser to the draw-er?  For now I used global variables.  Slap my
wrists, but I couldn't see an easy way to work it in otherwise.

- I added code to the vor drawing routine so that if the vor frequency
matches that of the nav1 or nav2 radio from FlightGear, I draw the
"to" and "from" radials (currently to = green, from = red).

Issue #2: I tried to use the AtlasPallette mechanism to define the
colors, but this didn't seem to work.  It appears that the existing
nav overlay colors are hardcoded and ignore the pallette settings.
Something's a little screwy there, but I hardcoded my red and green
colors along side the other hard coded colors and it worked.

Issue #3: Only the VOR's that are currently on screen are drawn.  That
means as soon as a tuned in VOR goes off the Atlas screen, the radial
lines disappear as well.  I'm not sure the best way to handle this,
but something needs to be thought about here ...

Issue #4: Given the projection you use (i.e. lat/lon to on-screen
pixel mapping) If I draw a straight line from a VOR out 50 miles, will
the end point of the line actually be 50 miles out at that radial, or
would we actually need some sort of "curved" line to accurately
represent the "straight" vor radial on screen?