Applied patch sent by Robert A. Knop <rknop@pobox.com>
(1) To OutputPS, add a "showpage" at the bottom so you can actually print
this out, and add a couple of "newpath" commands to remove stray lines.

(2) To MapPS, initialize nav*_freq.  (I believe these are used when Atlas
is talking to FlightGear, to show your raidal to the VOR station you're
tuned to.)

(3) To MapPS, OutputPS, and MapMaker, add some flags to the "features"
bitmap that let you specify VOR, NDB, and FIX navaids individually.  By
default, if you ask for navaids, you get them all (so that hopefully the
changes to MapMaker won't affect the behavior of Atlas).  MapPS has a
command line switch "--only-navaids" that lets you select which of these
you actually want to see.