The process_binary_file() routine didn't look at "local" normals for objects.  Instead, it always assumed that the vertices vector, v, and the normals vector, n, would be one-to-one, or that vertices would only used the first normal (n[0]).  This was wrong (I think - documentation on the file format is pretty well non-existant).
Sometimes an object (like a triangle strip, for example), will define,
in addition to a bunch of vertex indices, a bunch of normal indices.
Other times, the normal indices will be empty, in which case a vertex
index can also be used as an index into the global normals vector.
Clear?  I thought not.

I also cleaned up the new elevation cutting code.  It should now be
human-readable.