First Windows build in a long time - 20200525
And using CMake to generate build file of your choice.

This included several NEW files, some to support CMake -
  CMakeLists.txt
  CMakeModules/FindPLIB.cmake
  config.h.cmake
  projects/msvc/winsrc/getopt.c - replacing prev. getopt.cxx
  slimgear/simgear/misc/sgversionhelpers.hxx - prev. omitted
  slimgear/simgear/simgear-config.cmake

And modification of a considerable number of src files. The main fix was
including a "config.h" before all others, and for some reason replacing
min/max in a few others.